// src/app/modules/vehiculos/vehiculo/models/vehiculo-input.model.ts
export interface VehiculoInput {
    id?: number;
    modeloId?: number;
    tipoVehiculoId?: number;
    chapa?: string;
    color?: string;
    anho?: number;
    documentacion?: boolean;
    refrigerado?: boolean;
    nuevo?: boolean;
    fechaAdquisicion?: string;
    primerKilometraje?: number;
    capacidadKg?: number;
    capacidadPasajeros?: number;
    imagenesVehiculo?: string;
    imagenesDocumentos?: string;
    usuarioId?: number;
}